% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_B0}
\alias{generate_B0}
\title{Generates true sparse regression coefficients matrix B0 according to the settings in Section 5.1 of the paper.}
\usage{
  generate_B0(p, s, response_types)
}
\arguments{
  \item{p}{number of features}
  
  \item{s}{number of nonzero (signficant) features}
  
  \item{response_types}{vector of q response types. The entries of the vector must be one of the following: "continuous", "binary", or "count".}
  
}

\value{A p-by-q ground truth sparse regression coefficients matrix B0 with s(<=p) nonzero rows.
}

\description{
  This function generates a p-by-q ground truth sparse regression coefficients matrix B0 with s(<=p) nonzero rows. The simulation settings are those given in Section 5.1 of the manuscript. 
}
\examples{
n <- 100  # sample size 
p <- 20   # number of covariates
s <- 5    # number of significant covariates
q <- 3    # number of multiple responses 
response_types <- c('binary','continuous','count')

# Set seed to reproduce results later
set.seed(1)

# Generate p-by-q regression coefficients matrix B0, where s=# of nonzero rows.
B0 <- generate_B0(p, s, response_types)
B0
}