% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_Sigma0}
\alias{generate_Sigma0}
\title{Generates true covariance matrix Sigma0 according to the settings in Section 5.1 of the paper.}
\usage{
  generate_Sigma0(q, rho=0.5, sigma2=1)
}
\arguments{
  \item{q}{number of response variables}
  
  \item{rho}{autocorrelation parameter rho. Should be between 0 and 1.}
  
  \item{sigma2}{global variance parameter. Should be greater than 0.}
  
}

\value{A q-by-q ground truth covariance matrix Sigma0.
}

\description{
This function generates a q-by-q ground truth covariance matrix Sigma0. The simulation settings are those given in Section 5.1 of the manuscript. 
}

\examples{
q <- 3    # number of response variables 

# Generate true q-by-q covariance matrix Sigma0
Sigma0 <- generate_Sigma0(q)
}