% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_X}
\alias{generate_X}
\title{Generates n-by-p design matrix of covariates X according to the settings in Section 5.1 of the paper.}
\usage{
  generate_X(n, p, rho=0.5, sigma2=1)
}
\arguments{
  \item{n}{number of samples}
  
  \item{p}{number of covariates}
  
  \item{rho}{autocorrelation parameter. Should be between -1 and 1.}
  
  \item{sigma2}{global variance parameter. Should be greater than 0.}
  
}

\value{An n-by-p design matrix of covariates X.
}

\description{
This function generates an n-by-p design matrix of covariates. The simulation settings are those given in Section 5.1 of the manuscript. 
}

\examples{
n <- 100  # sample size 
p <- 20   # number of covariates

# Generate n-by-p design matrix X
X <- generate_X(n, p)
}